refactor: Inline landing page
There's no need to separate the contents of the landing page from
its current route.
